Update to GLM 0.9.7.4 Include GLM in src tree instead of using system framework

+++ src/raytracer.cpp
@@ -87,7 +87,7 @@ void Raytracer::run() const
             packetResult += rayColor;
         }
 
-        mImgBuffer->commit(*sample, packetResult / Sampler::sSamplesPerPixel);
+        mImgBuffer->commit(*sample, packetResult / (float)Sampler::sSamplesPerPixel);
     }
 }
